Rename filter-input to trips-utils.
