+++ frontend/src/components/ViewTrips/trip-utils.js
